Create PlayerCombat
C# script controlling the player's combat
holds information like the player's bullet's sprite, placing the crosshair where the mouse is on screen, and checking to see what weapon you are currently using and what keys will change to what weapon. The script also controls which direction to shoot the projectiles.